#106, #105, #104, Logic Balanced till can but upgrades, win loose popup, refactoring

+++ scripts/make_tables.sql
@@ -8,7 +8,7 @@ CONNECT 'jdbc:derby:../data/tables;create=true';
 CREATE TABLE regions (
     name VARCHAR(15),
     taxRate REAL,
-    entryCost REAL,
+    entryCost INT,
     logisticsCost INT,
     marketingCost INT,
     efficiencyCost INT,
